requirements.txt
python-telegram-bot==22.2
Flask==3.0.3
requests==2.32.4
openpyxl==3.1.5
pytesseract==0.3.13
pillow==11.3.0
pyTelegramBotAPI==4.27.0